pyproject.toml
[tool.poetry]
name = "kimetsuno"
version = "0.1.0"
description = ""
authors = ["insmnia <teamspirit3209@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.75.0"
uvicorn = "^0.17.6"
python-dotenv = "^0.19.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
motor = "^2.5.1"
bson = "^0.5.10"
python-multipart = "^0.0.5"
loguru = "^0.6.0"
pytest = "^7.1.1"
Faker = "^13.3.3"
factory-boy = "^3.2.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
